连接测试完成
